Rename "api" folder to "apis"
