+++ src/modules/options/OptionsApp.tsx
@@ -1,4 +1,4 @@
-import { TraktSettings } from '@api/TraktSettings';
+import { TraktSettings } from '@apis/TraktSettings';
 import { BrowserStorage, ServiceValue, StorageValuesOptions } from '@common/BrowserStorage';
 import { Errors } from '@common/Errors';
 import { EventDispatcher, OptionsChangeData, ServiceOptionsChangeData } from '@common/Events';
