+++ src/services/amazon-prime/AmazonPrimeApi.ts
@@ -1,5 +1,5 @@
 import { AmazonPrimeService } from '@/amazon-prime/AmazonPrimeService';
-import { ServiceApi } from '@api/ServiceApi';
+import { ServiceApi } from '@apis/ServiceApi';
 import { Errors } from '@common/Errors';
 import { RequestException, Requests } from '@common/Requests';
 import { ScriptInjector } from '@common/ScriptInjector';
