+++ src/services/nrk/NrkApi.ts
@@ -1,5 +1,5 @@
 import { NrkService } from '@/nrk/NrkService';
-import { ServiceApi } from '@api/ServiceApi';
+import { ServiceApi } from '@apis/ServiceApi';
 import { Requests } from '@common/Requests';
 import { IItem, Item } from '@models/Item';
 import * as moment from 'moment';
